liblzma-rs: remove compiler-trivial as c_uint casts (batch 1)

+++ liblzma-rs/src/common/alone_decoder.rs
@@ -209,12 +209,12 @@ pub type lzma_init_function = Option<
 >;
 pub const __DARWIN_NULL: *mut c_void = ::core::ptr::null_mut::<c_void>();
 pub const NULL: *mut c_void = __DARWIN_NULL;
-pub const UINT32_MAX: c_uint = 4294967295 as c_uint;
+pub const UINT32_MAX: c_uint = 4294967295;
 pub const UINT64_MAX: c_ulonglong = 18446744073709551615 as c_ulonglong;
 pub const true_0: c_int = 1 as c_int;
 pub const LZMA_VLI_UNKNOWN: c_ulonglong = UINT64_MAX;
 pub const LZMA_FILTER_LZMA1EXT: c_ulonglong = 0x4000000000000002 as c_ulonglong;
-pub const LZMA_LZMA1EXT_ALLOW_EOPM: c_uint = 0x1 as c_uint;
+pub const LZMA_LZMA1EXT_ALLOW_EOPM: c_uint = 0x1;
 pub const LZMA_MEMUSAGE_BASE: c_ulonglong = (1 as c_ulonglong) << 15 as c_int;
 unsafe extern "C" fn alone_decode(
     mut coder_ptr: *mut c_void,
@@ -229,10 +229,10 @@ unsafe extern "C" fn alone_decode(
 ) -> lzma_ret {
     let mut coder: *mut lzma_alone_coder = coder_ptr as *mut lzma_alone_coder;
     while *out_pos < out_size
-        && ((*coder).sequence as c_uint == SEQ_CODE as c_uint || *in_pos < in_size)
+        && ((*coder).sequence == SEQ_CODE || *in_pos < in_size)
     {
         let mut current_block_42: u64;
-        match (*coder).sequence as c_uint {
+        match (*coder).sequence {
             0 => {
                 if lzma_lzma_lclppb_decode(
                     &raw mut (*coder).options,
@@ -346,7 +346,7 @@ unsafe extern "C" fn alone_decode(
                     allocator,
                     &raw mut filters as *mut lzma_filter_info,
                 ) as lzma_ret;
-                if ret_ as c_uint != LZMA_OK as c_uint {
+                if ret_ != LZMA_OK {
                     return ret_;
                 }
                 (*coder).sequence = SEQ_CODE;
@@ -495,7 +495,7 @@ pub unsafe extern "C" fn lzma_alone_decoder(
     mut memlimit: u64,
 ) -> lzma_ret {
     let ret_: lzma_ret = lzma_strm_init(strm) as lzma_ret;
-    if ret_ as c_uint != LZMA_OK as c_uint {
+    if ret_ != LZMA_OK {
         return ret_;
     }
     let ret__0: lzma_ret = lzma_alone_decoder_init(
@@ -504,7 +504,7 @@ pub unsafe extern "C" fn lzma_alone_decoder(
         memlimit,
         0 as c_int != 0,
     ) as lzma_ret;
-    if ret__0 as c_uint != LZMA_OK as c_uint {
+    if ret__0 != LZMA_OK {
         lzma_end(strm);
         return ret__0;
     }
