+++ liblzma-rs/src/common/block_encoder.rs
@@ -263,7 +263,7 @@ unsafe extern "C" fn block_encode(
     }
     's_142: {
         let mut current_block_34: u64;
-        match (*coder).sequence as c_uint {
+        match (*coder).sequence {
             0 => {
                 let in_start: size_t = *in_pos;
                 let out_start: size_t = *out_pos;
@@ -297,8 +297,8 @@ unsafe extern "C" fn block_encode(
                         in_used,
                     );
                 }
-                if ret as c_uint != LZMA_STREAM_END as c_uint
-                    || action as c_uint == LZMA_SYNC_FLUSH as c_uint
+                if ret != LZMA_STREAM_END
+                    || action == LZMA_SYNC_FLUSH
                 {
                     return ret;
                 }
@@ -327,7 +327,7 @@ unsafe extern "C" fn block_encode(
                     *out_pos = (*out_pos).wrapping_add(1);
                     (*coder).compressed_size = (*coder).compressed_size.wrapping_add(1);
                 }
-                if (*(*coder).block).check as c_uint == LZMA_CHECK_NONE as c_uint {
+                if (*(*coder).block).check == LZMA_CHECK_NONE {
                     return LZMA_STREAM_END;
                 }
                 lzma_check_finish(&raw mut (*coder).check, (*(*coder).block).check);
@@ -371,7 +371,7 @@ unsafe extern "C" fn block_encoder_update(
     mut reversed_filters: *const lzma_filter,
 ) -> lzma_ret {
     let mut coder: *mut lzma_block_coder = coder_ptr as *mut lzma_block_coder;
-    if (*coder).sequence as c_uint != SEQ_CODE as c_uint {
+    if (*coder).sequence != SEQ_CODE {
         return LZMA_PROG_ERROR;
     }
     return lzma_next_filter_update(&raw mut (*coder).next, allocator, reversed_filters);
@@ -425,7 +425,7 @@ pub unsafe extern "C" fn lzma_block_encoder_init(
     if (*block).version > 1 as u32 {
         return LZMA_OPTIONS_ERROR;
     }
-    if (*block).check as c_uint > LZMA_CHECK_ID_MAX as c_uint {
+    if (*block).check > LZMA_CHECK_ID_MAX as c_uint {
         return LZMA_PROG_ERROR;
     }
     if lzma_check_is_supported((*block).check) == 0 {
@@ -502,13 +502,13 @@ pub unsafe extern "C" fn lzma_block_encoder(
     mut block: *mut lzma_block,
 ) -> lzma_ret {
     let ret_: lzma_ret = lzma_strm_init(strm) as lzma_ret;
-    if ret_ as c_uint != LZMA_OK as c_uint {
+    if ret_ != LZMA_OK {
         return ret_;
     }
     let ret__0: lzma_ret =
         lzma_block_encoder_init(&raw mut (*(*strm).internal).next, (*strm).allocator, block)
             as lzma_ret;
-    if ret__0 as c_uint != LZMA_OK as c_uint {
+    if ret__0 != LZMA_OK {
         lzma_end(strm);
         return ret__0;
     }
