+++ liblzma-rs/src/common/block_header_decoder.rs
@@ -131,7 +131,7 @@ pub unsafe extern "C" fn lzma_block_header_decode(
         .wrapping_add(1 as u32)
         .wrapping_mul(4 as u32)
         != (*block).header_size
-        || (*block).check as c_uint > LZMA_CHECK_ID_MAX as c_uint
+        || (*block).check > LZMA_CHECK_ID_MAX as c_uint
     {
         return LZMA_PROG_ERROR;
     }
@@ -151,7 +151,7 @@ pub unsafe extern "C" fn lzma_block_header_decode(
             &raw mut in_pos,
             in_size,
         ) as lzma_ret;
-        if ret_ as c_uint != LZMA_OK as c_uint {
+        if ret_ != LZMA_OK {
             return ret_;
         }
         if lzma_block_unpadded_size(block) == 0 as lzma_vli {
@@ -168,14 +168,14 @@ pub unsafe extern "C" fn lzma_block_header_decode(
             &raw mut in_pos,
             in_size,
         ) as lzma_ret;
-        if ret__0 as c_uint != LZMA_OK as c_uint {
+        if ret__0 != LZMA_OK {
             return ret__0;
         }
     } else {
         (*block).uncompressed_size = LZMA_VLI_UNKNOWN as lzma_vli;
     }
     let filter_count: size_t =
-        (*in_0.offset(1 as isize) as c_uint & 3 as c_uint).wrapping_add(1 as c_uint) as size_t;
+        (*in_0.offset(1 as isize) as c_uint & 3).wrapping_add(1) as size_t;
     let mut i_0: size_t = 0 as size_t;
     while i_0 < filter_count {
         let ret: lzma_ret = lzma_filter_flags_decode(
@@ -185,7 +185,7 @@ pub unsafe extern "C" fn lzma_block_header_decode(
             &raw mut in_pos,
             in_size,
         ) as lzma_ret;
-        if ret as c_uint != LZMA_OK as c_uint {
+        if ret != LZMA_OK {
             lzma_filters_free((*block).filters, allocator);
             return ret;
         }
