+++ liblzma-rs/src/common/filter_encoder.rs
@@ -558,7 +558,7 @@ pub unsafe extern "C" fn lzma_raw_encoder(
     mut filters: *const lzma_filter,
 ) -> lzma_ret {
     let ret_: lzma_ret = lzma_strm_init(strm) as lzma_ret;
-    if ret_ as c_uint != LZMA_OK as c_uint {
+    if ret_ != LZMA_OK {
         return ret_;
     }
     let ret__0: lzma_ret = lzma_raw_coder_init(
@@ -568,7 +568,7 @@ pub unsafe extern "C" fn lzma_raw_encoder(
         Some(coder_find as unsafe extern "C" fn(lzma_vli) -> *const lzma_filter_coder),
         1 as c_int != 0,
     ) as lzma_ret;
-    if ret__0 as c_uint != LZMA_OK as c_uint {
+    if ret__0 != LZMA_OK {
         lzma_end(strm);
         return ret__0;
     }
